% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_create.R
\name{api_create_concept}
\alias{api_create_concept}
\title{api_create_concept}
\usage{
api_create_concept(data, api_client)
}
\arguments{
\item{data}{Formatted concept data.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Creates concept using formatted data from uploaded yaml file.
}
