% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_formatter.R
\name{api_format_concept}
\alias{api_format_concept}
\title{api_format_concept}
\usage{
api_format_concept(concepts.data, phenotype.data)
}
\arguments{
\item{concepts.data}{List of unformatted concept data}

\item{phenotype.data}{Formatted parent phenotype data}
}
\value{
Formatted concept data (can return concept ID if concept type is
  existing_phenotype)
}
\description{
Formats concept data to standard accepted by the Concept Library API.
}
