% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_formatter.R
\name{api_format_phenotype}
\alias{api_format_phenotype}
\title{api_format_phenotype}
\usage{
api_format_phenotype(phenotype.data)
}
\arguments{
\item{phenotype.data}{Unformatted phenotype data}
}
\value{
Formatted phenotype data
}
\description{
Formats phenotype data to standard accepted by the Concept Library API.
}
