% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_validation.R
\name{api_validate_phenotype}
\alias{api_validate_phenotype}
\title{api_validate_phenotype}
\usage{
api_validate_phenotype(data, api_client, is.update)
}
\arguments{
\item{data}{Phenotype and concept data read from yaml file}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function.}

\item{is.update}{Whether the phenotype and concepts are being updated}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validates phenotype data and then validates concept and datasources
}
