% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{check_HTTP_response}
\alias{check_HTTP_response}
\title{check_HTTP_response}
\usage{
check_HTTP_response(response)
}
\arguments{
\item{response}{the crul::HTTPResponse object to check}
}
\description{
If the response status code is not 200, this function will stop execution and print the HTTP status that
was returned from the request.
}
\details{
If the status code is 200, the function will not return anything.
}
