% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{check_working_set_connection}
\alias{check_working_set_connection}
\title{check_working_set_connection}
\usage{
check_working_set_connection(api_client)
}
\arguments{
\item{api_client}{The API client to check.}
}
\description{
Working Sets API require an authenticated connection. This function throws an error if a public connection is used.
}
