% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_get_api_call.r
\name{cl_get_api_call}
\alias{cl_get_api_call}
\title{cl_get_api_call}
\usage{
cl_get_api_call(url, endpoint, auth)
}
\arguments{
\item{url}{URL of the concept library application}

\item{auth}{A crul::auth object. Will prompt for credentials if not supplied.}

\item{id}{A vector of one or more concept IDs to return}
}
\value{
A data frame with id and codes
}
\description{
Internal function to call the CL API to get data. Wrapped by functions that get concepts/working sets.
}
\examples{
get_concept(url='',id=c(90,91,4),user='username', password='*password*')
}
