% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_query_list}
\alias{clean_query_list}
\title{clean_query_list}
\usage{
clean_query_list(query_list)
}
\arguments{
\item{query_list}{The query parameter list to clean.}
}
\value{
The cleaned list of query parameters.
}
\description{
Removes elements that are NA or FALSE from the list, and sets elements that are TRUE to 1.
}
\details{
This function is used to create a cleaned query parameter list to pass into the query parameter of
crul::HTTPClient$get().
}
