% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{Collections}
\alias{Collections}
\title{Collections}
\description{
Collections object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of collections/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Collections}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Collections-get}{\code{Collections$get()}}
\item \href{#method-Collections-get_detail}{\code{Collections$get_detail()}}
\item \href{#method-Collections-clone}{\code{Collections$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-get"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-get}{}}}
\subsection{Method \code{get()}}{
Queries collections/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries collections/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$get_detail(collection_id)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
