% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptLibraryClient.R
\docType{package}
\name{ConceptLibraryClient}
\alias{ConceptLibraryClient}
\title{ConceptLibraryClient: A Concept Library API client.}
\description{
This package acts as a client for the Concept Library. Each function represents an HTTP request to the API to
interact with Phenotypes, Concepts, Data Sources, Working Sets and Tags. Before using these functions, the package
user must create an HttpClient object by calling one of the connection functions. Once this object is created, it
must passed into any API function that is called.
}
\section{Authenticated API connection}{

By calling \code{\link{connect_to_API}}, the user must provide login credentials either in the function itself, or in
the popup window if none are given.

This api client can then be used with any of the API functions available. The public API can still be accessed even
with an authenticated connection.
}

\section{Public API connection}{

\code{\link{connect_to_public_API}} allows for a connection without requiring authentication. This connection object
can only be used with functions which allow use of the public API. These will either be functions which explicity get
published data, e.g. \code{\link{get_published_concepts}}, or functions which have the \code{use_public_api} argument
available, which must be set to \code{TRUE}.

Functions which fit neither of the above criteria are simply not available to the public API.
}

