% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{Concepts}
\alias{Concepts}
\title{Concepts}
\description{
Concepts object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of concepts/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Concepts}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Concepts-get}{\code{Concepts$get()}}
\item \href{#method-Concepts-get_versions}{\code{Concepts$get_versions()}}
\item \href{#method-Concepts-get_detail}{\code{Concepts$get_detail()}}
\item \href{#method-Concepts-get_codelist}{\code{Concepts$get_codelist()}}
\item \href{#method-Concepts-clone}{\code{Concepts$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Concepts-get"></a>}}
\if{latex}{\out{\hypertarget{method-Concepts-get}{}}}
\subsection{Method \code{get()}}{
Queries concepts/, with optional query parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Concepts$get(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Concepts-get_versions"></a>}}
\if{latex}{\out{\hypertarget{method-Concepts-get_versions}{}}}
\subsection{Method \code{get_versions()}}{
Queries concepts/{id}/get-versions/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Concepts$get_versions(concept_id)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Concepts-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Concepts-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries concepts/{id}/detail/ or concepts/{id}/version/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Concepts$get_detail(concept_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Concepts-get_codelist"></a>}}
\if{latex}{\out{\hypertarget{method-Concepts-get_codelist}{}}}
\subsection{Method \code{get_codelist()}}{
Queries concepts/{id}/export/codes/ or
 concepts/{id}/version/{id}/export/codes/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Concepts$get_codelist(concept_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Concepts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Concepts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Concepts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
