% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{connect_to_API}
\alias{connect_to_API}
\title{connect_to_API}
\usage{
connect_to_API(user = NA, password = NA, url = API_URL, public = TRUE)
}
\arguments{
\item{user}{A username (optional)}

\item{password}{The user's password (optional)}

\item{url}{The URL to connect to (optional). Defaults to main Concept Library website.}

\item{public}{When TRUE connects to the public API (authentication not required). When set to FALSE, a username and
password are required to connect to the authenticated API. Default is TRUE.}
}
\value{
crul::HttpClient
}
\description{
This function creates a crul::HttpClient object from a url and optionally a user and password. If the user or
password are not provided and public is set to FALSE, then a login box is created to enter this information.
}
