% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{connect_to_public_API}
\alias{connect_to_public_API}
\title{connect_to_public_API}
\usage{
connect_to_public_API()
}
\arguments{
\item{url}{The Concept Library API url}
}
\value{
crul::HttpClient
}
\description{
This function creates crul::HttpClient object from a url. This connection does not use authentication and can only
be used for public API access.
}
