% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Connection}
\alias{Connection}
\title{Connection}
\description{
Connection object, storing HttpClient connection and allowing access to
 endpoint instances, e.g. templates, phenotypes
}
\examples{
# Authenticated login, with url set to HDRUK
client = ConceptLibraryClient::Connection$new(
  url=ConceptLibraryClient::DOMAINS$HDRUK
)

# Authenticated login, with custom URL
client = ConceptLibraryClient::Connection$new(
  url='custom-url.com/'
)

# Non-authenticated connection
client = ConceptLibraryClient::Connection$new(public=FALSE)

# Accessing phenotypes
phenotype_detail = client$phenotypes$get('PH1', version_id=2)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{HttpClient}}{(crul::HttpClient) Connection object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{HttpClient}}{(crul::HttpClient) Connection object}

\item{\code{templates}}{(ConceptLibraryClient::Templates) Templates instance}

\item{\code{phenotypes}}{(ConceptLibraryClient::Phenotypes) Phenotypes instance}

\item{\code{concepts}}{(ConceptLibraryClient::Concepts) Concepts instance}

\item{\code{collections}}{(ConceptLibraryClient::Collections) Collections instance}

\item{\code{tags}}{(ConceptLibraryClient::Tags) Tags instance}

\item{\code{datasources}}{(ConceptLibraryClient::Datasources) Datasources instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientConnection-new}{\code{Connection$new()}}
\item \href{#method-ClientConnection-clone}{\code{Connection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConnection-new}{}}}
\subsection{Method \code{new()}}{
Create a connection to the ConceptLibrary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$new(
  username = NA,
  password = NA,
  public = FALSE,
  url = DEFAULT_CONNECTION_URL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{(string) Username for authentication}

\item{\code{password}}{(string) Password for authentication}

\item{\code{public}}{(bool) Flag to determine whether login is required}

\item{\code{url}}{(string) URL used when connecting to the ConceptLibrary}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
