% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_auth.r
\name{create_auth}
\alias{create_auth}
\title{create_auth}
\usage{
create_auth(user = NA, password = NA)
}
\arguments{
\item{user}{User name for authentication (if username and password left blank, will pop up a login box)}

\item{password}{Password for authentication}
}
\value{
A crul::auth object for use in API calls.
}
\description{
Function to handle authentication
}
\examples{
credentials = create_auth(user='dan', password='fido123');
credentials = create_auth();  #popup a login box for user name and password.
credentials = create_auth(user='dan');  #popup a login box with user name prepopulated with dan
}
