% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources.R
\name{Datasources}
\alias{Datasources}
\title{Datasources}
\description{
Datasources object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of data-sources/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Datasources}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Datasources-get}{\code{Datasources$get()}}
\item \href{#method-Datasources-get_detail}{\code{Datasources$get_detail()}}
\item \href{#method-Datasources-clone}{\code{Datasources$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datasources-get"></a>}}
\if{latex}{\out{\hypertarget{method-Datasources-get}{}}}
\subsection{Method \code{get()}}{
Queries data-sources/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datasources$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datasources-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Datasources-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries data-sources/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datasources$get_detail(datasource_id)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datasources-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Datasources-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datasources$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
