% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DOMAINS}
\alias{DOMAINS}
\title{DOMAINS}
\format{
An object of class \code{list} of length 5.
}
\usage{
DOMAINS
}
\description{
List containing default domains for connecting to the Concept Library
}
\keyword{datasets}
