% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{Endpoint}
\alias{Endpoint}
\title{Endpoint}
\value{
Formatted URL

Returns response after validating

Response object
}
\description{
Endpoint object, storing HTTP client object and handling requests and
 responses to API
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{STATUS_CODE_SUCCESS}}{(list) Contains successful response codes}

\item{\code{HttpClient}}{(crul::HttpClient) Connection object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{STATUS_CODE_SUCCESS}}{(list) Contains successful response codes}

\item{\code{HttpClient}}{(crul::HttpClient) Connection object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Endpoint-new}{\code{Endpoint$new()}}
\item \href{#method-Endpoint-clone}{\code{Endpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-new}{}}}
\subsection{Method \code{new()}}{
Create an endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$new(connection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{(crul::HttpClient) Connection object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
