% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection_by_id}
\alias{get_collection_by_id}
\title{get_collection_by_id}
\usage{
get_collection_by_id(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The collection's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the collection.
}
\description{
Lists a collection by id.
}
\examples{
get_collection_by_id('20')

api_client = connect_to_API(public = FALSE)
get_collection_by_id('20', api_client = api_client)

}
