% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collections}
\alias{get_collections}
\title{get_collections}
\usage{
get_collections(api_client = connect_to_API())
}
\arguments{
\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the collections.
}
\description{
Lists all available collections.
}
\examples{
get_collections()

api_client = connect_to_API(public = FALSE)
get_collections(api_client = api_client)

}
