% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept.r
\name{get_concept}
\alias{get_concept}
\title{get_concept}
\usage{
get_concept(url, id, user = NA, password = NA)
}
\arguments{
\item{url}{URL of the concept library application}

\item{id}{A vector of one or more concept IDs to return}

\item{user}{User name for authentication (if username and password left blank, will pop up a login box)}

\item{password}{Password for authentication}
}
\value{
A data frame with id and codes
}
\description{
Retrieve a concept or set of concepts from Concept Library
}
\examples{
get_concept(url='',id=c(90,91,4),user='username', password='*password*')
}
