% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_by_id}
\alias{get_concept_by_id}
\title{get_concept_by_id}
\usage{
get_concept_by_id(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The concept's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the concept.
}
\description{
Lists a concept by id.
}
\examples{
get_concept_by_id(api_client, 'C47')
get_concept_by_id(api_client, 'C47', use_public_api = TRUE)

}
