% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_code_list}
\alias{get_concept_code_list}
\title{get_concept_code_list}
\usage{
get_concept_code_list(api_client, id)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function.}

\item{id}{The concept's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the code list.
}
\description{
Exports the code list of the latest version of a concept for the user. This cannot be used with the public API.
}
\examples{
get_concept_code_list(api_client, 'C47')

}
