% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_code_list_by_version}
\alias{get_concept_code_list_by_version}
\title{get_concept_code_list_by_version}
\usage{
get_concept_code_list_by_version(id, version_id, api_client = connect_to_API())
}
\arguments{
\item{id}{The concept's id.}

\item{version_id}{The concept version's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the code list.
}
\description{
Exports the code list of a specific version of a concept.
}
\examples{
get_concept_code_list_by_version('C714', '2567')

api_client = connect_to_API(public = FALSE)
get_concept_code_list_by_version('C714', '2567', api_client = api_client)

}
