% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_codes.r
\name{get_concept_codes}
\alias{get_concept_codes}
\title{get_concept_codes}
\usage{
get_concept_codes(url, id, version = NA, auth = NA)
}
\arguments{
\item{url}{URL of the concept library application}

\item{id}{A vector of one or more concept IDs to return}

\item{auth}{A crul::auth object (use create_auth to create); if blank will popup asking for authentication}

\item{id}{A vector defining the version of each concept to return (if the entire vector is NA, get the latest version of all; if individual items are NA, get the latest version of that concept, but specify versions for others)}
}
\value{
A dataframe with id and codes
}
\description{
Retrieve a concept or set of concepts from Concept Library
}
\examples{
get_concept_codes(url='',id=c(90,91,4),user='username', password='*password*')
}
