% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_detail}
\alias{get_concept_detail}
\title{get_concept_detail}
\usage{
get_concept_detail(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The concept's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the concept detail.
}
\description{
Lists the concept detail of the latest version (or latest published version if using public API).
}
\examples{
get_concept_detail(api_client, 'C47')
get_concept_detail(api_client, 'C47', use_public_api = TRUE)

}
