% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concepts}
\alias{get_concepts}
\title{get_concepts}
\usage{
get_concepts(
  api_client = connect_to_API(),
  search = NA,
  tag_ids = NA,
  show_only_my_concepts = FALSE,
  show_deleted_concepts = FALSE,
  show_only_validated_concepts = FALSE,
  brand = NA,
  author = NA,
  owner_username = NA,
  do_not_show_versions = FALSE,
  must_have_published_versions = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}

\item{search}{Search by part of concept name (do not put wild characters here)}

\item{tag_ids}{Specify vector of tags ids (get tags from get_tags())}

\item{show_only_my_concepts}{Only show concepts owned by me. Default is FALSE. Can't be used with public API.}

\item{show_deleted_concepts}{Also show deleted concepts. Default is FALSE. Can't be used with public API.}

\item{show_only_validated_concepts}{Show only validated concepts. Default is FALSE.}

\item{brand}{Show only concepts with a specified brand.}

\item{author}{Search by part of the author name.}

\item{owner_username}{Search by full username of the owner. Can't be used with public API.}

\item{do_not_show_versions}{Do not show concepts versions. Default is FALSE (versions are shown).}

\item{must_have_published_versions}{Show only concepts which have a published version. Default is FALSE. Can't be
used with public API.}
}
\value{
A dataframe containing the concepts matching the query.
}
\description{
Lists all available concepts for the user.
}
\examples{
get_concepts()

api_client = connect_to_API(public = FALSE)
get_concept(api_client = api_client)
get_concepts(
  api_client = api_client,
  search = 'Alcohol',
  tag_ids = c(11,4),
  show_only_my_concepts = TRUE,
  show_deleted_concepts = TRUE,
  show_only_validated_concepts = TRUE,
  brand = 'HDRUK',
  author = 'Kuan',
  owner_username = 'a.john',
  do_not_show_versions = TRUE,
  must_have_published_versions = TRUE)

}
