% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{get_data_source_by_id}
\alias{get_data_source_by_id}
\title{get_data_source_by_id}
\usage{
get_data_source_by_id(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The data source's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the data source.
}
\description{
Lists a data source by id and the associated phenotypes.
}
\examples{
get_data_source_by_id('26')

api_client = connect_to_API(public = FALSE)
get_data_souce_by_id('26', api_client = api_client)

}
