% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{get_data_sources}
\alias{get_data_sources}
\title{get_data_sources}
\usage{
get_data_sources(api_client, search = NA, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{search}{Search by part of the data source name.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the data sources.
}
\description{
Lists all data sources and the phenotypes associated with each.
}
\examples{
get_data_sources(api_client, search = 'hospital')
get_data_sources(api_client, use_public_api = TRUE)

}
