% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_path}
\alias{get_full_path}
\title{get_full_path}
\usage{
get_full_path(path, use_public_api)
}
\arguments{
\item{path}{The path to prepend to.}

\item{use_public_api}{If the public API should be used instead.}
}
\value{
The full API path
}
\description{
Simple function which prepends api/ or api/public to the beginning of a path, depending on whether the public api
should be used.
}
