% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_by_id}
\alias{get_phenotype_by_id}
\title{get_phenotype_by_id}
\usage{
get_phenotype_by_id(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The phenotype's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the phenotype.
}
\description{
Lists a phenotype by id and the data sources associated with it.
}
\examples{
get_phenotype_by_id(api_client, 'PH3183')
get_phenotype_by_id(api_client, 'PH3183', use_public_api = TRUE)

}
