% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_code_list_by_version}
\alias{get_phenotype_code_list_by_version}
\title{get_phenotype_code_list}
\usage{
get_phenotype_code_list_by_version(
  id,
  version_id,
  api_client = connect_to_API()
)
}
\arguments{
\item{id}{The phenotype's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function.}
}
\value{
A dataframe containing the code list.
}
\description{
Exports the code list of a specific version of a phenotype.
}
\examples{
get_phenotype_code_list('PH1', '2', client)

api_client = connect_to_API(public = FALSE)
get_phenotype_code_list('PH1', '2', api_client = api_client)

}
