% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_detail}
\alias{get_phenotype_detail}
\title{get_phenotype_detail}
\usage{
get_phenotype_detail(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The phenotype's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the phenotype detail.
}
\description{
Lists the phenotype detail of the latest version (or latest published version if using public API).
}
\examples{
get_phenotype_detail('PH1')

api_client = connect_to_API(public = FALSE)
get_phenotype_detail('PH1', api_client = api_client)

}
