% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_detail_by_version}
\alias{get_phenotype_detail_by_version}
\title{get_phenotype_detail_by_version}
\usage{
get_phenotype_detail_by_version(
  api_client,
  id,
  version_id,
  use_public_api = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The phenotype's id.}

\item{version_id}{The phenotype version's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the phenotype detail.
}
\description{
Lists the phenotype detail of the specified version.
}
\examples{
get_phenotype_detail_by_version(api_client, 'PH3183', '10865')
get_phenotype_detail_by_version(api_client, 'PH3183', '10865', use_public_api = TRUE)

}
