% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_versions}
\alias{get_phenotype_versions}
\title{get_phenotype_versions}
\usage{
get_phenotype_versions(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The phenotype's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the phenotype's versions.
}
\description{
Lists all the versions of the phenotype
}
\examples{
get_phenotype_versions(api_client, 'PH3183')
get_phenotype_versions(api_client, 'PH3183', use_public_api = TRUE)

}
