% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_published_concepts}
\alias{get_published_concepts}
\title{get_published_concepts}
\usage{
get_published_concepts(
  api_client,
  search = NA,
  tag_ids = NA,
  show_only_validated_concepts = FALSE,
  brand = NA,
  author = NA,
  do_not_show_versions = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{search}{Search by part of concept name (do not put wild characters here)}

\item{tag_ids}{Specify vector of tags ids (get tags from get_tags())}

\item{show_only_validated_concepts}{Show only validated concepts. Default is FALSE.}

\item{brand}{Show only concepts with a specified brand.}

\item{author}{Search by part of the author name.}

\item{do_not_show_versions}{Do not show concepts versions. Default is FALSE (versions are shown).}
}
\value{
A dataframe containing the concepts matching the query.
}
\description{
Lists the published concepts.
}
\examples{
get_published_concepts(api_client)
get_published_concepts(
  api_client,
  search = 'Alcohol',
  tag_ids = c(11,4),
  show_only_validate_concepts = TRUE,
  brand = 'HDRUK',
  author = 'Kuan',
  do_not_show_versions = TRUE)

}
