% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_published_phenotypes}
\alias{get_published_phenotypes}
\title{get_published_phenotypes}
\usage{
get_published_phenotypes(
  api_client,
  search = NA,
  tag_ids = NA,
  show_only_validated_phenotypes = FALSE,
  brand = NA,
  author = NA,
  do_not_show_versions = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions}

\item{search}{Search by part of phenotype name (do not put wild characters here)}

\item{tag_ids}{Specify vector of tags ids (get tags from get_tags())}

\item{show_only_validated_phenotypes}{Show only validated phenotypes. Default is FALSE.}

\item{brand}{Show only phenotypes with a specified brand.}

\item{author}{Search by part of the author name.}

\item{do_not_show_versions}{Do not show phenotypes versions. Default is FALSE (versions are shown).}
}
\value{
A dataframe containing the published phenotypes matching the query.
}
\description{
Lists the published phenotypes and the data sources associated with each.
}
\examples{
getPublishedPhenotypes(api_client)
getPublishedPhenotypes(
  api_client,
  search = 'Alcohol',
  tag_ids = c(11, 4),
  show_only_validated_phenotypes = TRUE,
  brand = 'HDRUK',
  author = 'Kuan',
  do_not_show_versions = TRUE)

}
