% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tag_by_id}
\alias{get_tag_by_id}
\title{get_tag_by_id}
\usage{
get_tag_by_id(api_client, id)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The tag's id.}
}
\value{
A dataframe containing the tag.
}
\description{
Lists a tag by id. A public API connection can be used here.
}
\examples{
get_tag_by_id(api_client, '123')

}
