% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_working_set.r
\name{get_working_set}
\alias{get_working_set}
\title{get_working_set}
\usage{
get_working_set(url, id, user = NA, password = NA)
}
\arguments{
\item{url}{URL of the concept library application}

\item{id}{A vector of one or more concept IDs to return}

\item{user}{User name for authentication (if username and password left blank, will pop up a login box)}

\item{password}{Password for authentication}
}
\value{
A data frame with working set contents
}
\description{
Retrieve a single working set from the Concept Library
}
\examples{
get_working_set(url='',id=1,user='username', password='*password*')
}
