% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_set_by_id}
\alias{get_working_set_by_id}
\title{get_working_set_by_id}
\usage{
get_working_set_by_id(api_client, id)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function.}

\item{id}{The working set's id.}
}
\value{
A dataframe containing the working set.
}
\description{
Lists a working set by id for the user. This cannot be used with the public API.
}
\examples{
get_working_set_by_id(api_client, '123')

}
