% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_set_code_list}
\alias{get_working_set_code_list}
\title{get_working_set_code_list}
\usage{
get_working_set_code_list(api_client, id)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function. This must be an authenticated connection.}

\item{id}{The working set's id.}
}
\value{
A dataframe containing the working set's code list.
}
\description{
Exports the code list of the latest version of a working set for the user. This cannot be used with the public
API.
}
\examples{
api_client = connect_to_API(public = FALSE)
get_working_set_code_list(api_client, '123')

}
