% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_set_code_list_by_version}
\alias{get_working_set_code_list_by_version}
\title{get_working_set_code_list_by_version}
\usage{
get_working_set_code_list_by_version(api_client, id, version_id)
}
\arguments{
\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. This must be an authenticated
connection.}

\item{id}{The working set's id.}

\item{version_id}{The working set version's id.}
}
\value{
A dataframe containing the working set's code list.
}
\description{
Exports the code list of a specific version of a working set for the user. This cannot be used with the public
API.
}
\examples{
api_client = connect_to_API(public = FALSE)
get_working_set_code_list_by_version(api_client, '123', '456')

}
