% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_set_detail}
\alias{get_working_set_detail}
\title{get_working_set_detail}
\usage{
get_working_set_detail(api_client, id, do_not_show_codes = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function. This must be an authenticated connection.}

\item{id}{The working set's id.}

\item{do_not_show_codes}{Do not show working set codes. Default is FALSE (codes are shown).}
}
\value{
A dataframe containing the working set detail.
}
\description{
Lists the working set detail of the latest version. This cannot be used with the public API.
}
\examples{
api_client = connect_to_API(public = FALSE)
get_working_set_detail(api_client, '123')
get_working_set_detail(api_client, '123', do_not_show_codes = TRUE)

}
