% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_set_versions}
\alias{get_working_set_versions}
\title{get_working_set_versions}
\usage{
get_working_set_versions(api_client, id)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function.}

\item{id}{The working set's id.}
}
\value{
A dataframe containing the working set's versions.
}
\description{
Lists all the versions of the working set for the user. This cannot be used with the public API.
}
\examples{
get_working_set_versions(api_client, '123')

}
