% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_sets.R
\name{get_working_sets}
\alias{get_working_sets}
\title{get_working_sets}
\usage{
get_working_sets(
  api_client,
  search = NA,
  tag_ids = NA,
  show_only_my_workingsets = FALSE,
  show_deleted_workingsets = FALSE,
  brand = NA,
  author = NA,
  owner_username = NA,
  do_not_show_versions = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function.}

\item{search}{Search by part of working set name (do not put wild characters here)}

\item{tag_ids}{Specify vector of tags ids (get tags from get_tags())}

\item{show_only_my_workingsets}{Only show working sets owned by me. Default is FALSE.}

\item{show_deleted_workingsets}{Also show deleted working sets. Default is FALSE.}

\item{brand}{Only show working sets with specified brand.}

\item{author}{Search by part of the author name.}

\item{owner_username}{Search by full username of the owner.}

\item{do_not_show_versions}{Do not show working sets versions. Default is FALSE (versions are shown).}
}
\value{
A dataframe containing the working sets.
}
\description{
Lists all available working sets for the user. This cannot be used with the public API.
}
\examples{
get_working_sets(api_client)
get_working_sets(
  search = 'diab',
  tag_ids = c(11,4),
  show_only_my_workingsets = TRUE,
  show_deleted_workingsets = TRUE,
  brand = 'ADP',
  author = 'moh',
  owner_username = 'a.john',
  do_not_show_versions = TRUE)

}
