% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{getLogin}
\alias{getLogin}
\title{getLogin}
\usage{
getLogin(userName = "")
}
\arguments{
\item{userName}{A user name (optional). If provided, the login box will be prepopulated with this user name.}
}
\description{
This function pops up a login box for username and password, with masking of the password field. It returns a vector
where the first item is the username and the second is the password. Clear this as soon as it is used, to avoid
password existing in your environment.
}
\details{
This can be used for any purpose. If you want to connect to SAIL, try SAILConnect instead, which
uses this function and also connects to PR_SAIL and returns a database connection.

Note that the login box may appear behind your active window.
}
\examples{
details = getLogin();
}
\keyword{SAILDBUtils}
