% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_connection_authenticated}
\alias{is_connection_authenticated}
\title{is_connection_public}
\usage{
is_connection_authenticated(api_client)
}
\arguments{
\item{api_client}{The api connection to check.}
}
\value{
TRUE if authenticated, FALSE if public.
}
\description{
Determine whether the api connection is authenticated (TRUE) or public (FALSE).
}
