% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{is_empty}
\usage{
is_empty(input)
}
\arguments{
\item{input}{(any) Input to be compared}
}
\value{
TRUE if empty, FALSE otherwise
}
\description{
Checks whether given input is empty
}
