% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{Phenotypes}
\alias{Phenotypes}
\title{Phenotypes}
\value{
Response object

Response object

Response object

Response object

Response object

Response object

Data read in from the file

Formatted phenotype data

Formatted concept data

Formatted concept data

Concept component

Formatted phenotype data
}
\description{
Phenotypes object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of phenotypes/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Phenotypes}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ALLOWED_FILE_EXTENSIONS}}{(list) List of file extensions allowed}

\item{\code{PHENOTYPE_IGNORE_FILEDS}}{(list) List of fields to ignore}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ALLOWED_FILE_EXTENSIONS}}{(list) List of file extensions allowed}

\item{\code{PHENOTYPE_IGNORE_FILEDS}}{(list) List of fields to ignore}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Phenotypes-get}{\code{Phenotypes$get()}}
\item \href{#method-Phenotypes-get_versions}{\code{Phenotypes$get_versions()}}
\item \href{#method-Phenotypes-get_detail}{\code{Phenotypes$get_detail()}}
\item \href{#method-Phenotypes-get_codelist}{\code{Phenotypes$get_codelist()}}
\item \href{#method-Phenotypes-save_definition_file}{\code{Phenotypes$save_definition_file()}}
\item \href{#method-Phenotypes-create}{\code{Phenotypes$create()}}
\item \href{#method-Phenotypes-update}{\code{Phenotypes$update()}}
\item \href{#method-Phenotypes-clone}{\code{Phenotypes$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-get"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-get}{}}}
\subsection{Method \code{get()}}{
Queries phenotypes/, with optional query parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$get(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(list) List of optional parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-get_versions"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-get_versions}{}}}
\subsection{Method \code{get_versions()}}{
Queries phenotypes/{id}/get-versions/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$get_versions(phenotype_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotype_id}}{(string) Id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries phenotypes/{id}/detail/ or phenotypes/{id}/version/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$get_detail(phenotype_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotype_id}}{(string) Id of entity to query}

\item{\code{version_id}}{(integer) Version id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-get_codelist"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-get_codelist}{}}}
\subsection{Method \code{get_codelist()}}{
Queries phenotypes/{id}/export/codes/ or
 phenotypes/{id}/version/{id}/export/codes/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$get_codelist(phenotype_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotype_id}}{(string) Id of entity to query}

\item{\code{version_id}}{(integer) Version id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-save_definition_file"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-save_definition_file}{}}}
\subsection{Method \code{save_definition_file()}}{
Formats phenotype detail and saves it to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$save_definition_file(path, phenotype_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(string) Path to save the file}

\item{\code{phenotype_id}}{(string) Id of entity to query}

\item{\code{version_id}}{(integer) Version id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-create"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-create}{}}}
\subsection{Method \code{create()}}{
Creates a new phenotype based on the .yaml file supplied
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$create(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(string) Path to definition file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-update"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-update}{}}}
\subsection{Method \code{update()}}{
Updates an existing phenotype with details from the .yaml file supplied
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$update(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(string) Path to definition file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotypes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotypes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotypes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
