% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file}
\alias{read_file}
\title{read_file}
\usage{
read_file(file.path)
}
\arguments{
\item{file.path}{Path including file name}
}
\value{
If input passes validation, returns contents of file,
        returns NA otherwise.
}
\description{
Validates input and reads in contents of specified file.
}
