% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{save_phenotype_definition}
\alias{save_phenotype_definition}
\title{save_phenotype_definition}
\usage{
save_phenotype_definition(
  dir,
  id,
  version_id = NA,
  api_client = connect_to_API()
)
}
\arguments{
\item{id}{The phenotype's id.}

\item{version_id}{The phenotype version's id, defaults to NA. Leave as default if using public api.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the phenotype's data
}
\description{
Saves the Phenotype YAML definition file locally
}
\examples{
save_phenotype_definition('C:/path/to/folder', 'definition-file.yaml', 'PH1')

api_client = connect_to_API()
save_phenotype_definition('C:/path/to/folder', 'definition-file.yaml', 'PH1', '2', api_client)

}
