% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_gateway_proxy}
\alias{set_gateway_proxy}
\title{set_gateway_proxy}
\usage{
set_gateway_proxy()
}
\value{
NA
}
\description{
Sets the user system environment variable to the gateway proxy.
This method is only called when the user is utilising the DOMAINS$GATEWAY URL
}
