% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_write_field}
\alias{should_write_field}
\title{should_write_field}
\usage{
should_write_field(value)
}
\arguments{
\item{value}{}
}
\value{
TRUE if content, FALSE otherwise
}
\description{
Determines whether the field has content that should be written
}
