% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{Tags}
\alias{Tags}
\title{Tags}
\description{
Tags object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of tags/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Tags}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tags-get}{\code{Tags$get()}}
\item \href{#method-Tags-get_detail}{\code{Tags$get_detail()}}
\item \href{#method-Tags-clone}{\code{Tags$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tags-get"></a>}}
\if{latex}{\out{\hypertarget{method-Tags-get}{}}}
\subsection{Method \code{get()}}{
Queries tags/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tags$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tags-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Tags-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries tags/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tags$get_detail(tag_id)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tags-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tags-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tags$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
