% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{Templates}
\alias{Templates}
\title{Templates}
\value{
Response object

Response object

Response object
}
\description{
Templates object, inheriting from ConceptLibraryClient::Endpoint - allows
 querying of templates/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Templates}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Templates-get}{\code{Templates$get()}}
\item \href{#method-Templates-get_versions}{\code{Templates$get_versions()}}
\item \href{#method-Templates-get_detail}{\code{Templates$get_detail()}}
\item \href{#method-Templates-clone}{\code{Templates$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Templates-get"></a>}}
\if{latex}{\out{\hypertarget{method-Templates-get}{}}}
\subsection{Method \code{get()}}{
Queries templates/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Templates$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Templates-get_versions"></a>}}
\if{latex}{\out{\hypertarget{method-Templates-get_versions}{}}}
\subsection{Method \code{get_versions()}}{
Queries templates/{id}/get-versions/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Templates$get_versions(template_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_id}}{(string) Id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Templates-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Templates-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries templates/{id}/detail/ or templates/{id}/version/{id}/detail/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Templates$get_detail(template_id, version_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_id}}{(string) Id of entity to query}

\item{\code{version_id}}{(integer) Version id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Templates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Templates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Templates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
