% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_parse_doi}
\alias{try_parse_doi}
\title{try_parse_doi}
\usage{
try_parse_doi(publications)
}
\arguments{
\item{publications}{(list) List of publications from entity definition file}
}
\value{
List containing detail and doi
}
\description{
Tries to parse all DOIs from list of publications
}
