% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_create.R
\name{update_phenotype}
\alias{update_phenotype}
\title{update_phenotype
upload_phenotype}
\usage{
update_phenotype(file.path, api_client)
}
\arguments{
\item{file.path}{File path pointing to yaml file containing phenotype and concept data}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function}
}
\value{
List of updated, formatted phenotype and concept data
}
\description{
Updates existing phenotypes and concepts based on provided yaml file.
}
