% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_create.R
\name{upload_phenotype}
\alias{upload_phenotype}
\title{upload_phenotype}
\usage{
upload_phenotype(file.path, api_client)
}
\arguments{
\item{file.path}{File path pointing to yaml file containing phenotype and concept data}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function}
}
\value{
List of formatted phenotype and concept data, including uploaded ids
}
\description{
Reads yaml file containing phenotype and concept data, validates, formats and uploads to the Concept Library.
}
