% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_codes}
\alias{validate_codes}
\title{validate_codes}
\usage{
validate_codes(codes)
}
\arguments{
\item{codes}{List or string containing codes.}
}
\value{
TRUE if valid codelist, FALSE otherwise
}
\description{
Validates list or string containing concept codelist.
}
