% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_csv}
\alias{validate_csv}
\title{validate_csv}
\usage{
validate_csv(file.path)
}
\arguments{
\item{file.path}{Path including file name}
}
\value{
TRUE if valid file and structure, FALSE otherwise
}
\description{
Validates csv file, including structure.
}
