% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_type}
\alias{validate_type}
\title{validate_type}
\usage{
validate_type(input, type)
}
\arguments{
\item{input}{(any) Input to be compared}

\item{type}{(string) Type in {'string', 'number', 'bool'}}
}
\value{
TRUE if type matches, FALSE otherwise
}
\description{
Validates input against a type; string, number, bool
}
