% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_yaml_file}
\alias{write_yaml_file}
\title{write_yaml_file}
\usage{
write_yaml_file(yaml_data, phenotype.id, concept.ids, file.path, api_client)
}
\arguments{
\item{phenotype.id}{Uploaded phenotype's id}

\item{concept.ids}{Uploaded phenotype's concept ids}

\item{file.path}{Path including file name}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function.}
}
\description{
Writes modified yaml data to file.
}
