% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptLibraryClient.R
\docType{package}
\name{ConceptLibraryClient}
\alias{ConceptLibraryClient}
\title{ConceptLibraryClient: A Concept Library API Client}
\description{
This package acts as a client for the ConceptLibrary API
}
\section{Authenticated API connection}{

By calling \code{ConceptLibraryClient::Connection$new()}, the user must
provide login credentials either in the function itself, or in the popup
window if none are given.
}

\section{Non-authenticated API connection}{

A non-authenticated, or public, connection can be made to the ConceptLibrary
by calling \code{ConceptLibraryClient::Connection$new(public=TRUE)}.
}

\section{Modifying API connection URL}{

The connection function also provides a \code{url} parameter to specify a
different URL to send the API requests to.

We provide default connection urls through:
\code{ConceptLibraryClient::DOMAINS}
}

\section{Making API requests}{

The following functions can be used to retrieve the various types of data
stored in the Concept Library.
\subsection{Phenotypes}{
\itemize{
\item \code{client$phenotypes$get}
\item \code{client$phenotypes$get_versions}
\item \code{client$phenotypes$get_detail}
\item \code{client$phenotypes$get_codelist}
\item \code{client$phenotypes$save_definition_file}
\item \code{client$phenotypes$create}
\item \code{client$phenotypes$update}
}
}

\subsection{Concepts}{
\itemize{
\item \code{client$concepts$get}
\item \code{client$concepts$get_versions}
\item \code{client$concepts$get_detail}
\item \code{client$concepts$get_codelist}
}
}

\subsection{Templates}{
\itemize{
\item \code{client$templates$get}
\item \code{client$templates$get_versions}
\item \code{client$templates$get_detail}
}
}

\subsection{Collections}{
\itemize{
\item \code{client$templates$get}
\item \code{client$templates$get_detail}
}
}

\subsection{Tags}{
\itemize{
\item \code{client$templates$get}
\item \code{client$templates$get_detail}
}
}

\subsection{Datasources}{
\itemize{
\item \code{client$templates$get}
\item \code{client$templates$get_detail}
}
}
}

