% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{connect_to_API}
\alias{connect_to_API}
\title{connect_to_API}
\usage{
connect_to_API(user = NA, password = NA)
}
\arguments{
\item{user}{A username (optional)}

\item{password}{The user's password (optional)}

\item{url}{The Concept Library API url}
}
\value{
crul::HttpClient
}
\description{
This function creates a crul::HttpClient object from a url and optionally a user and password. If the user or
password are not provided, then a login box is created to enter this information.
}
