% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_by_id}
\alias{get_concept_by_id}
\title{get_concept_by_id}
\usage{
get_concept_by_id(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The concept's id.}

\item{api_client}{The HttpClient returned by the connect_to_API function. Optional, a public API connection is
created if left blank.}
}
\value{
A dataframe containing the concept.
}
\description{
Lists a concept by id.
}
\examples{
get_concept_by_id('C714')

api_client = connect_to_API(public = FALSE)
get_concept_by_id('C714', api_client = api_client)

}
