% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_code_list}
\alias{get_concept_code_list}
\title{get_concept_code_list}
\usage{
get_concept_code_list(id, api_client)
}
\arguments{
\item{id}{The concept's id.}

\item{api_client}{The HttpClient returned by the connectToAPI function. This must be an authenticated connection.}
}
\value{
A dataframe containing the code list.
}
\description{
Exports the code list of the latest version of a concept for the user. This cannot be used with the public API.
}
\examples{
api_client = connect_to_API(public = FALSE)
get_concept_code_list('C714', api_client)

}
