% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_code_list_by_version}
\alias{get_concept_code_list_by_version}
\title{get_concept_code_list_by_version}
\usage{
get_concept_code_list_by_version(
  api_client,
  id,
  version_id,
  use_public_api = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The concept's id.}

\item{version_id}{The concept version's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the code list.
}
\description{
Exports the code list of a specific version of a concept.
}
\examples{
get_concept_code_list_by_version(api_client, 'C47', '54903')
get_concept_code_list_by_version(api_client, 'C47', '24798', use_public_api = TRUE)

}
