% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_detail}
\alias{get_concept_detail}
\title{get_concept_detail}
\usage{
get_concept_detail(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The concept's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the concept detail.
}
\description{
Lists the concept detail of the latest version (or latest published version if using public API).
}
\examples{
get_concept_detail('C714')

api_client = connect_to_API(public = FALSE)
get_concept_detail('C714', api_client = api_client)

}
