% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{get_concept_versions}
\alias{get_concept_versions}
\title{get_concept_versions}
\usage{
get_concept_versions(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The concept's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the concept versions.
}
\description{
Lists all the versions of the concept.
}
\examples{
get_concept_versions('C714')

api_client = connect_to_API(public = FALSE)
get_concept_versions('C714', api_client = api_client)

}
