% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{get_data_source_by_id}
\alias{get_data_source_by_id}
\title{get_data_source_by_id}
\usage{
get_data_source_by_id(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The data source's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the data source.
}
\description{
Lists a data source by id and the associated phenotypes.
}
\examples{
get_data_source_by_id(api_client, '425')
get_data_souce_by_id(api_client, '425', use_public_api = TRUE)

}
