% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{get_data_sources}
\alias{get_data_sources}
\title{get_data_sources}
\usage{
get_data_sources(api_client = connect_to_API(), search = NA)
}
\arguments{
\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}

\item{search}{Search by part of the data source name.}
}
\value{
A dataframe containing the data sources.
}
\description{
Lists all data sources and the phenotypes associated with each.
}
\examples{
get_data_sources()

api_client = connect_to_API(public = FALSE)
get_data_sources(api_client = api_client, search = 'hospital')

}
