% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_path}
\alias{get_full_path}
\title{get_full_path}
\usage{
get_full_path(path, api_client)
}
\arguments{
\item{path}{The path to prepend to.}

\item{api_client}{The api_client to determine connection type}
}
\value{
The full API path
}
\description{
Simple function which prepends api/v1 or api/v1/public to the beginning of a path, depending on whether the api
client's connection type (authenticated or public)
}
