% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_by_id}
\alias{get_phenotype_by_id}
\title{get_phenotype_by_id}
\usage{
get_phenotype_by_id(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The phenotype's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the phenotype.
}
\description{
Lists a phenotype by id and the data sources associated with it.
}
\examples{
get_phenotype_by_id('PH1')

api_client = connect_to_API(public = FALSE)
get_phenotype_by_id('PH1', api_client = api_client)

}
