% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_detail}
\alias{get_phenotype_detail}
\title{get_phenotype_detail}
\usage{
get_phenotype_detail(api_client, id, use_public_api = FALSE)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI or connectToPublicAPI functions.}

\item{id}{The phenotype's id.}

\item{use_public_api}{If the public API should be accessed instead. Default is FALSE.}
}
\value{
A dataframe containing the phenotype detail.
}
\description{
Lists the phenotype detail of the latest version (or latest published version if using public API).
}
\examples{
get_phenotype_detail(api_client, 'PH3183')
get_phenotype_detail(api_client, 'PH3183', use_public_api = TRUE)

}
