% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_detail_by_version}
\alias{get_phenotype_detail_by_version}
\title{get_phenotype_detail_by_version}
\usage{
get_phenotype_detail_by_version(id, version_id, api_client = connect_to_API())
}
\arguments{
\item{id}{The phenotype's id.}

\item{version_id}{The phenotype version's id.}

\item{api_client}{The HttpClient returned by the connect_to_API function. Optional, a public API connection is
created if left blank.}
}
\value{
A dataframe containing the phenotype detail.
}
\description{
Lists the phenotype detail of the specified version.
}
\examples{
get_phenotype_detail_by_version('PH1', '2')

api_client = connect_to_API(public = FALSE)
get_phenotype_detail_by_version('PH1', '2', api_client = api_client)

}
