% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_versions}
\alias{get_phenotype_versions}
\title{get_phenotype_versions}
\usage{
get_phenotype_versions(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The phenotype's id.}

\item{api_client}{The HttpClient returned by the \code{\link{connect_to_API}} function. Optional, a public API
connection is created if left blank.}
}
\value{
A dataframe containing the phenotype's versions.
}
\description{
Lists all the versions of the phenotype
}
\examples{
get_phenotype_versions('PH1')

api_client = connect_to_API()
get_phenotype_versions('PH1', api_client = api_client)

}
