% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{get_phenotypes}
\usage{
get_phenotypes(
  api_client,
  search = NA,
  tag_ids = NA,
  show_only_my_phenotypes = FALSE,
  show_deleted_phenotypes = FALSE,
  show_only_validated_phenotypes = FALSE,
  brand = NA,
  author = NA,
  owner_username = NA,
  do_not_show_versions = FALSE,
  must_have_published_versions = FALSE
)
}
\arguments{
\item{api_client}{The HttpClient returned by the connectToAPI function.}

\item{search}{Search by part of phenotype name (do not put wild characters here)}

\item{tag_ids}{Specify vector of tags ids (get tags from get_tags())}

\item{show_only_my_phenotypes}{Only show phenotypes owned by me. Default is FALSE.}

\item{show_deleted_phenotypes}{Also show deleted phenotypes. Default is FALSE.}

\item{show_only_validated_phenotypes}{Show only validated phenotypes. Default is FALSE.}

\item{brand}{Show only phenotypes with a specified brand.}

\item{author}{Search by part of the author name.}

\item{owner_username}{Search by full username of the owner.}

\item{do_not_show_versions}{Do not show phenotypes versions. Default is FALSE (versions are shown).}

\item{must_have_published_versions}{Show only phenotypes which have a published version. Default is FALSE.}
}
\value{
A dataframe containing the phenotypes matching the query.
}
\description{
Lists all available phenotypes for the user and the data sources associate with each.
}
\examples{
getPhenotypes(api_client)
getPhenotypes(
  api_client,
  search = 'Alcohol',
  tag_ids = c(11,4),
  show_only_my_phenotypes = TRUE,
  show_deleted_phenotypes = TRUE,
  show_only_validated_phenotypes = TRUE,
  brand = 'HDRUK',
  author = 'Kuan',
  owner_username = 'a.john',
  do_not_show_versions = TRUE,
  must_have_published_versions = TRUE)

}
