% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tag_by_id}
\alias{get_tag_by_id}
\title{get_tag_by_id}
\usage{
get_tag_by_id(id, api_client = connect_to_API())
}
\arguments{
\item{id}{The tag's id.}

\item{api_client}{The HttpClient returned by the connect_to_API function. Optional, a public API connection is
created if left blank.}
}
\value{
A dataframe containing the tag.
}
\description{
Lists a tag by id. A public API connection can be used here.
}
\examples{
get_tag_by_id('20')

api_client = connect_to_API(public = FALSE)
get_tag_by_id('20', api_client = api_client)

}
