% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{get_tags}
\title{get_tags}
\usage{
get_tags(api_client = connect_to_API())
}
\arguments{
\item{api_client}{The HttpClient returned by the connect_to_API function. Optional, a public API connection is
created if left blank.}
}
\value{
A dataframe containing the tags.
}
\description{
Lists all available tags. A public API connection can be used here.
}
\examples{
get_tags()

api_client = connect_to_API(public = FALSE)
get_tags(api_client = api_client)

}
