% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_working_set.r
\name{get_working_set}
\alias{get_working_set}
\title{get_working_set}
\usage{
get_working_set(url, id, version = NA, auth = NA)
}
\arguments{
\item{url}{URL of the concept library application}

\item{id}{A working set ID to return}

\item{version}{Version of working set to return (NA = latest version)}

\item{auth}{Credentials for API call (create with create_auth() or leave blank to be prompted)}
}
\value{
A data frame with working set contents
}
\description{
Retrieve a single working set from the Concept Library
}
\examples{
get_working_set(url='https://conceptlibrary.saildatabank.com',id=1,auth=credentials)
}
