% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_type}
\alias{validate_type}
\title{validate_type}
\usage{
validate_type(input, data.type)
}
\arguments{
\item{input}{The input data}

\item{data.type}{The expected data type of input}
}
\value{
TRUE/FALSE dependent on assertion
}
\description{
Simple function which validates input to expected data type.
}
